Extract main loop out of terminal.cpp
